requirements.txt
flask==3.0.0
openai==1.12.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
numpy==1.26.4
Werkzeug==3.0.1
typing-extensions==4.9.0

